Updated CayneMap
created a copy of the Take3 dancer for Cayne, replaced the model in cayne's map, and put the laser pointer texture on there. I don't know what he changed with the animation, so he'll have to re-do those actions. Sorry! We don't want to be changing files that other people may be using.